+++ df_core/peer.h
@@ -59,7 +59,7 @@ private:
 
 public:
     peer(long id, std::string ip, int pss_port, int data_port, double position, std::shared_ptr<spdlog::logger> logger);
-    peer(long id, std::string ip, int pss_port, int data_port, double position, long pss_boot_time, int pss_view_size, long pss_sleep_interval, int pss_gossip_size,
+    peer(long id, std::string ip, int pss_port, int data_port, double position, long pss_boot_time, int pss_view_size, int pss_sleep_interval, int pss_gossip_size,
             int logging_interval, int anti_entropy_interval, std::string logging_dir, int rep_max, int rep_min, int max_age, bool local_message, int local_interval, float reply_chance, bool smart, std::shared_ptr<spdlog::logger> logger);
     void print_view();
     void start();
